BuildingManager
-BuildingManager now handles temperatures.
