+++ TeamProject/GlobalWarmingGame/GlobalWarmingGame.csproj
@@ -47,8 +47,8 @@
     <Compile Include="Action\Instruction.cs" />
     <Compile Include="Action\InstructionType.cs" />
     <Compile Include="Interactions\Interactables\Buildings\Building.cs" />
-    <Compile Include="Interactions\Interactables\Buildings\BuildingManager.cs" />
-    <Compile Include="Interactions\Interactables\IBuildable.cs" />
+    <Compile Include="BuildingManager.cs" />
+    <Compile Include="Interactions\Interactables\Buildings\IBuildable.cs" />
     <Compile Include="Interactions\Interactables\Buildings\WorkBench.cs" />
     <Compile Include="Interactions\Interactables\Environment\CoalNode.cs" />
     <Compile Include="Interactions\Interactables\Environment\StoneNode.cs" />
